added tests, pushing to see failures
