+++ demisto_sdk/commands/lint/tests/command_builder_test.py
@@ -29,12 +29,12 @@ def test_build_xsoar_linter_py3_command(files):
     """Build xsoar linter command"""
     from demisto_sdk.commands.lint.commands_builder import build_xsoar_linter_command
 
-    output = build_xsoar_linter_command(files, "base")
+    output = build_xsoar_linter_command(files, "base", all_packs=True)
     files = [str(file) for file in files]
     expected = (
         f"pylint --ignore=CommonServerPython.py,demistomock.py,CommonServerUserPython.py,"
         "conftest.py,.venv -E --disable=all --msg-template='{abspath}:{line}:{column}: {msg_id} {obj}: {msg}'"
-        " --enable=E9002,E9003,E9004,E9005,E9006,E9007,E9010,E9011,E9012,E9013,W9013, --load-plugins "
+        " --enable=E9002,E9003,E9004,E9005,E9006,E9007,E9010,E9011,E9012,W9013, --load-plugins "
         f"base_checker, {' '.join(files)}"
     )
     assert output == expected
@@ -45,12 +45,12 @@ def test_build_xsoar_linter_py2_command(files):
     """Build xsoar linter command"""
     from demisto_sdk.commands.lint.commands_builder import build_xsoar_linter_command
 
-    output = build_xsoar_linter_command(files, "base")
+    output = build_xsoar_linter_command(files, "base", all_packs=True)
     files = [str(file) for file in files]
     expected = (
         f"pylint --ignore=CommonServerPython.py,demistomock.py,CommonServerUserPython.py,"
         "conftest.py,.venv -E --disable=all --msg-template='{abspath}:{line}:{column}: {msg_id} {obj}: {msg}' "
-        "--enable=E9002,E9003,E9004,E9005,E9006,E9007,E9010,E9011,E9012,E9013,W9013, --load-plugins "
+        "--enable=E9002,E9003,E9004,E9005,E9006,E9007,E9010,E9011,E9012,W9013, --load-plugins "
         f"base_checker, {' '.join(files)}"
     )
     assert output == expected
